+++ src/contact/__test__/updatedContact.trigger.test.ts
@@ -57,8 +57,7 @@ describe('update contact trigger', () => {
     )
   })
 
-  // TODO This feature is currently not implemented
-  xit('should fetch contacts sorted by update_at and generate deduplication based on passed trigger field',
+  it('should fetch contacts sorted by update_at and generate deduplication based on passed trigger field',
     async () => {
       const bundle = {
         inputData: {
@@ -84,8 +83,7 @@ describe('update contact trigger', () => {
       )
     })
 
-  // TODO This feature is currently not implemented
-  xit('should fetch companies sorted by update_at and generate deduplication based on passed trigger field',
+  it('should fetch companies sorted by update_at and generate deduplication based on passed trigger field',
     async () => {
       const bundle = {
         inputData: {