+++ src/deal/__test__/updatedDeal.trigger.test.ts
@@ -30,8 +30,7 @@ describe('deals update trigger', () => {
     )
   })
 
-  // TODO This feature is currently not implemented
-  xit('should fetch deals sorted by update_at and generate deduplication based on passed trigger field', async () => {
+  it('should fetch deals sorted by update_at and generate deduplication based on passed trigger field', async () => {
     const bundle = {
       inputData: {
         trigger_field: 'custom_fields.Very_Complicated.CustomField'