+++ src/lead/__test__/updatedLead.trigger.test.ts
@@ -30,8 +30,7 @@ describe('update lead trigger', () => {
     )
   })
 
-  // TODO This feature is currently not implemented
-  xit('should fetch leads sorted by update_at and generate deduplication based on passed trigger field', async () => {
+  it('should fetch leads sorted by update_at and generate deduplication based on passed trigger field', async () => {
     const bundle = {
       inputData: {
         trigger_field: 'custom_fields.Custom Field 1'